requirements.txt
azure-functions
numpy
joblib
sklearn